+++ ocaml-lsp-server/src/code_actions/code_action.ml
@@ -27,11 +27,4 @@ let source_text doc (loc : Loc.t) =
   String.sub (Msource.text source) ~pos:start ~len:(end_ - start)
 ;;
 
-let workspace_edit doc text_edits =
-  let uri = Document.uri doc in
-  let version = Document.version doc in
-  let textDocument = OptionalVersionedTextDocumentIdentifier.create ~uri ~version () in
-  let edits = List.map text_edits ~f:(fun edit -> `TextEdit edit) in
-  let edit = TextDocumentEdit.create ~textDocument ~edits in
-  WorkspaceEdit.create ~documentChanges:[ `TextDocumentEdit edit ] ()
-;;
+let workspace_edit = Document.edit